pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
databases = {extras = ["sqlite"], version = "^0.9.0"}
flake8 = "^7.0.0"
pytest = "^8.1.1"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
httpx = "^0.27.0"

[tool.poetry.scripts]
dev = "src.scripts.dev:main"
test = "src.scripts.test:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
